Move AddPet
